chore: feature-gate root prover and Bn254 in continuations
