--- crates/continuations/src/bn254.rs
@@ -1,133 +1,0 @@
-use std::array::from_fn;
-
-use num_bigint::BigUint;
-use openvm_stark_sdk::config::baby_bear_poseidon2::{DIGEST_SIZE, F};
-use openvm_verify_stark_host::pvs::DagCommit;
-use p3_bn254::Bn254;
-use p3_field::{PrimeCharacteristicRing, PrimeField, PrimeField32};
-
-pub const BN254_BYTES: usize = 32;
-
-/// Wrapper for an array of big-endian bytes, representing an unsigned big integer. Each commit can
-/// be converted to a Bn254 using the trivial identification as natural numbers or into a `u32`
-/// digest by decomposing the big integer base-`F::MODULUS`.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub struct CommitBytes([u8; BN254_BYTES]);
-
-impl CommitBytes {
-    pub fn new(bytes: [u8; BN254_BYTES]) -> Self {
-        Self(bytes)
-    }
-
-    pub fn as_slice(&self) -> &[u8; BN254_BYTES] {
-        &self.0
-    }
-
-    pub fn reverse(&mut self) {
-        self.0.reverse();
-    }
-}
-
-impl From<[F; DIGEST_SIZE]> for CommitBytes {
-    fn from(value: [F; DIGEST_SIZE]) -> Self {
-        Self::from(babybear_digest_to_bn254(&value))
-    }
-}
-
-impl From<[u32; DIGEST_SIZE]> for CommitBytes {
-    fn from(value: [u32; DIGEST_SIZE]) -> Self {
-        Self(u32_digest_to_bytes(&value))
-    }
-}
-
-impl From<Bn254> for CommitBytes {
-    fn from(value: Bn254) -> Self {
-        Self(bn254_to_bytes(value))
-    }
-}
-
-impl From<[Bn254; 1]> for CommitBytes {
-    fn from(value: [Bn254; 1]) -> Self {
-        CommitBytes::from(value[0])
-    }
-}
-
-impl<F: PrimeCharacteristicRing> From<CommitBytes> for [F; DIGEST_SIZE] {
-    fn from(value: CommitBytes) -> Self {
-        bytes_to_u32_digest(&value.0).map(F::from_u32)
-    }
-}
-
-impl From<CommitBytes> for Bn254 {
-    fn from(value: CommitBytes) -> Self {
-        bytes_to_bn254(&value.0)
-    }
-}
-
-fn babybear_digest_to_bn254(digest: &[F; DIGEST_SIZE]) -> Bn254 {
-    let mut ret = Bn254::ZERO;
-    let order = Bn254::from_u32(F::ORDER_U32);
-    let mut base = Bn254::ONE;
-    digest.iter().for_each(|&x| {
-        ret += base * Bn254::from_u32(x.as_canonical_u32());
-        base *= order;
-    });
-    ret
-}
-
-fn bytes_to_bn254(bytes: &[u8; BN254_BYTES]) -> Bn254 {
-    let order = Bn254::from_u32(1 << 8);
-    let mut ret = Bn254::ZERO;
-    let mut base = Bn254::ONE;
-    for byte in bytes.iter().rev() {
-        ret += base * Bn254::from_u8(*byte);
-        base *= order;
-    }
-    ret
-}
-
-fn bn254_to_bytes(bn254: Bn254) -> [u8; BN254_BYTES] {
-    let mut ret = [0u8; BN254_BYTES];
-    let bytes = bn254.as_canonical_biguint().to_bytes_be();
-    let start = BN254_BYTES - bytes.len();
-    ret[start..].copy_from_slice(&bytes);
-    ret
-}
-
-fn bytes_to_u32_digest(bytes: &[u8; BN254_BYTES]) -> [u32; DIGEST_SIZE] {
-    let mut bigint = BigUint::ZERO;
-    for byte in bytes.iter() {
-        bigint <<= 8;
-        bigint += BigUint::from(*byte);
-    }
-    let order = F::ORDER_U32;
-    from_fn(|_| {
-        let bigint_digit = bigint.clone() % order;
-        let digit = if bigint_digit == BigUint::ZERO {
-            0u32
-        } else {
-            bigint_digit.to_u32_digits()[0]
-        };
-        bigint /= order;
-        digit
-    })
-}
-
-fn u32_digest_to_bytes(digest: &[u32; DIGEST_SIZE]) -> [u8; BN254_BYTES] {
-    bn254_to_bytes(babybear_digest_to_bn254(&digest.map(F::from_u32)))
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub struct DagCommitBytes {
-    pub cached_commit: CommitBytes,
-    pub pre_hash: CommitBytes,
-}
-
-impl<F: PrimeCharacteristicRing> From<DagCommitBytes> for DagCommit<F> {
-    fn from(value: DagCommitBytes) -> Self {
-        DagCommit {
-            cached_commit: value.cached_commit.into(),
-            vk_pre_hash: value.pre_hash.into(),
-        }
-    }
-}
