+++ crates/continuations/src/circuit/root/verifier/air.rs
@@ -23,7 +23,6 @@ use p3_field::{Field, PrimeCharacteristicRing};
 use p3_matrix::Matrix;
 
 use crate::{
-    bn254::{CommitBytes, DagCommitBytes},
     circuit::{
         root::{
             bus::{
@@ -37,6 +36,7 @@ use crate::{
         CONSTRAINT_EVAL_CACHED_INDEX,
     },
     utils::{digests_to_poseidon2_input, pad_slice_to_poseidon2_input},
+    CommitBytes, DagCommitBytes,
 };
 
 #[repr(C)]
