+++ crates/continuations/src/prover/mod.rs
@@ -3,45 +3,52 @@ use openvm_cpu_backend::CpuBackend;
 use openvm_cuda_backend::{BabyBearBn254Poseidon2GpuEngine, GpuBackend};
 use openvm_recursion_circuit::system::VerifierSubCircuit;
 
+#[cfg(feature = "root-prover")]
+use crate::{circuit::root::RootTraceGenImpl, RootSC};
 use crate::{
     circuit::{
         deferral::{hook::DeferralHookTraceGenImpl, inner::DeferralInnerTraceGenImpl},
         inner::InnerTraceGenImpl,
-        root::RootTraceGenImpl,
     },
-    RootSC, SC,
+    SC,
 };
 
 mod deferral;
 mod inner;
-mod root;
 mod utils;
 
 pub use deferral::*;
 pub use inner::*;
-pub use root::*;
 pub use utils::*;
+
+#[cfg(feature = "root-prover")]
+mod root;
+#[cfg(feature = "root-prover")]
+pub use root::*;
 
 pub type InnerCpuProver<const MAX_NUM_PROOFS: usize> =
     InnerAggregationProver<CpuBackend<SC>, VerifierSubCircuit<MAX_NUM_PROOFS>, InnerTraceGenImpl>;
-pub type RootCpuProver = RootProver<CpuBackend<RootSC>, VerifierSubCircuit<1>, RootTraceGenImpl>;
-pub type DeferralInnerCpuProver =
-    DeferralInnerProver<CpuBackend<SC>, VerifierSubCircuit<2>, DeferralInnerTraceGenImpl>;
-pub type DeferralHookCpuProver =
-    DeferralHookProver<CpuBackend<SC>, VerifierSubCircuit<1>, DeferralHookTraceGenImpl>;
-
 #[cfg(feature = "cuda")]
 pub type InnerGpuProver<const MAX_NUM_PROOFS: usize> =
     InnerAggregationProver<GpuBackend, VerifierSubCircuit<MAX_NUM_PROOFS>, InnerTraceGenImpl>;
-#[cfg(feature = "cuda")]
+
+#[cfg(feature = "root-prover")]
+pub type RootCpuProver = RootProver<CpuBackend<RootSC>, VerifierSubCircuit<1>, RootTraceGenImpl>;
+#[cfg(all(feature = "cuda", feature = "root-prover"))]
 pub type RootGpuProver = RootProver<
     <BabyBearBn254Poseidon2GpuEngine as openvm_stark_backend::StarkEngine>::PB,
     VerifierSubCircuit<1>,
     RootTraceGenImpl,
 >;
+
+pub type DeferralInnerCpuProver =
+    DeferralInnerProver<CpuBackend<SC>, VerifierSubCircuit<2>, DeferralInnerTraceGenImpl>;
 #[cfg(feature = "cuda")]
 pub type DeferralInnerGpuProver =
     DeferralInnerProver<GpuBackend, VerifierSubCircuit<2>, DeferralInnerTraceGenImpl>;
+
+pub type DeferralHookCpuProver =
+    DeferralHookProver<CpuBackend<SC>, VerifierSubCircuit<1>, DeferralHookTraceGenImpl>;
 #[cfg(feature = "cuda")]
 pub type DeferralHookGpuProver =
     DeferralHookProver<GpuBackend, VerifierSubCircuit<1>, DeferralHookTraceGenImpl>;
